NATS Pull Reader to allow client to define a stream, subject and optional filter to read message from NATS
The configuration allows the client to define a batchSize that determines how many messages to pull from NATS thereby providing a  way to ensure we dont consume too much memory.  When NaTS has no messages it will return 404 to the consumer. This is can be checked by the client by calling the method NatsReader::noMoreMsgs, which returns true if NATS had no more matching messages on subjject and or filter name.

The unit test give good examples of usage.

Note: This is not a standalone executable jar application